implemented wav file test through mic and amp via bypassing DSP

+++ esp32_firmware/components/amp/amp.c
@@ -1,11 +1,12 @@
 #include "amp.h"
 #include <stdio.h>
+#include <inttypes.h>
 
 void amp_init(void)
 {
 }
 
-void amp_write_block(float *buffer, int size)
+void amp_write_block(int32_t *buffer, int size)
 {
-    printf("OUT: %f\n", buffer[0]);
+    printf("OUT: %" PRId32 "\n", buffer[0]);
 }
