+++ esp32_firmware/components/amp/amp.h
@@ -1,8 +1,9 @@
 #ifndef AMP_H
 #define AMP_H
+#include <inttypes.h>
+#include <stdint.h>
 
 void amp_init(void);
-
-void amp_write_block(float *buffer, int size);
+void amp_write_block(int32_t *buffer, int size);
 
 #endif
